MP3/requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.3
cachelib==0.9.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Caching==2.1.0
gunicorn==21.2.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
packaging==23.2
pluggy==1.3.0
protobuf==4.21.12
pycountry==22.3.5
pytest==7.4.3
python-dotenv==1.0.0
soupsieve==2.5
Werkzeug==3.0.1
